requirements.txt
Django==1.10.5
psycopg2==2.6.2
google-api-python-client==1.6.2
django-nopassword==3.0.1
gunicorn==19.6.0
whitenoise==3.3.0
